itemleri tek tek silmek için js işlemleri yapıldı

+++ src/js/index.js
@@ -4,18 +4,18 @@ TodoApp.items = [];
 
 
 
-//Bu alanda itemleri localstorage'tan getItem yöntemi ile alıp json.parse ederek TodoApp.items'e atadık
+
 if(localStorage.getItem('items')){
     TodoApp.items = JSON.parse(localStorage.getItem('items'));
 }
 
-//Bu alanda localstroage'ten setItem ile göndererek sürekli güncel bilgi gönderiyoruz
+
 function updateLocalStorage(){
     localStorage.setItem('items', JSON.stringify(TodoApp.items));
 }
 
 
-//Eğer task tanımsız ya da herhangi bir değeri yok geri döndür,değilse todoapp itemine id,name,completed,timespent ve timeractive ekle
+
 function addTask(task){
     if(task===undefined || task.lenght===0){
         return;
@@ -35,22 +35,20 @@ function addTask(task){
 //window.addTask = addTask;
 
 
-function toggleTask(task_id) {
-    //TodoApp itemlerinde filtreleme yapıyoruz t parametresindeki id ve task id eşit halini task_local e atıyoruz.
-    const task_local = TodoApp.items.filter(t => t.id === task_id);
-    //Eğer task_local de eşitlik var ise
+function toggleTask(task_id) {    
+    const task_local = TodoApp.items.filter(t => t.id === task_id);    
     if(task_local) {
         task_local[0].completed = !task_local[0].completed;
     }
-    //Eğer task_local de eşitlik yok ise completed edilmiştir ve timeractive false edilmiştir.
     if(task_local[0].completed){
         task_local[0].timerActive= false;
     }
     updateLocalStorage();
     updateUI();
 }
-
 TodoApp.toggleTask = toggleTask;
+
+
 
 //Delete ıtems
 function deleteTask(task_id) {
@@ -58,8 +56,10 @@ function deleteTask(task_id) {
     updateLocalStorage();
     updateUI();
 }
-
 TodoApp.deleteTask = deleteTask;
+
+
+
 
 //DELETE all task
 function deleteAllTasks(){
@@ -67,7 +67,6 @@ function deleteAllTasks(){
     updateLocalStorage();
     updateUI();
 }
-
 TodoApp.deleteAllTasks = deleteAllTasks;
 
 // Start Timer
@@ -81,11 +80,8 @@ function startTimer(task_id){
     updateLocalStorage();
     updateUI();
 }
-
 TodoApp.startTimer = startTimer;
 
-
-//timerActive
 
 // Pause Timer 
 function pauseTimer(task_id){
@@ -96,7 +92,6 @@ function pauseTimer(task_id){
     updateLocalStorage();
     updateUI();
 }
-
 TodoApp.pauseTimer = pauseTimer;
 
 
@@ -121,7 +116,6 @@ setInterval(function(){
 
 
 
-
 // UI ---------------------------
 
 const getTasksMarkup = (is_completed) => 
@@ -131,6 +125,7 @@ const getTasksMarkup = (is_completed) =>
         ${secondsToHms(task.timeSpent)} 
         ${task.completed ? '' : 
         `<button onclick="${task.timerActive ? 'TodoApp.pauseTimer(' + task.id + ')' : 'TodoApp.startTimer(' + task.id + ')'}">${task.timerActive ? 'Durdur' : 'Başlat'}</button>`}
+        ${task.completed ? '' : `<a onclick="TodoApp.deleteTaskUi(${task.id})" href="#" id="itemDelete" class="delete-all float-right">Task remove</a>`}
         </li>`).join('\n');
 
 //window.getTasksMarkup = getTasksMarkup;
@@ -161,7 +156,19 @@ form.addEventListener('submit', function(e){
 const deleteAll = document.getElementById('btnDeleteAll');
 
 deleteAll.addEventListener('click', function(e){
-    deleteAllTasks();
+    if(confirm('are you sure ?')){
+        deleteAllTasks();
+    }    
 });
 
-    
+
+
+// delete items ui
+function deleteTaskUi(task_id){
+    if(confirm('are you sure ?')){
+        deleteTask(task_id);
+    }
+}
+TodoApp.deleteTaskUi = deleteTaskUi;
+
+
